% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mode_ctp}
\alias{mode_ctp}
\title{Mode of the CTP distribution}
\usage{
mode_ctp(a, b, gama)
}
\arguments{
\item{a}{alpha parameter}

\item{b}{b parameter (b >= 0)}

\item{gama}{gamma parameter}
}
\value{
the mode (a single integer), or, in the exact tie case,
  a numeric vector of length 2 giving both consecutive modes
}
\description{
Mode of the CTP distribution
}
