% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mode_zictp}
\alias{mode_zictp}
\title{Mode of the ZI-CTP distribution}
\usage{
mode_zictp(a, b, gama, omega)
}
\arguments{
\item{a}{alpha parameter}

\item{b}{b parameter (b >= 0)}

\item{gama}{gamma parameter}

\item{omega}{zero-inflation parameter}
}
\value{
the mode (a single integer)
}
\description{
Mode of the ZI-CTP distribution
}
