% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctp.R
\name{pctp}
\alias{pctp}
\title{Cumulative Distribution Function of the CTP Distribution}
\usage{
pctp(q, a, b, gama, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{a}{Parameter a.}

\item{b}{Parameter b.}

\item{gama}{Parameter gamma.}

\item{lower.tail}{Logical; if TRUE, probabilities are P(X <= q),
otherwise P(X > q).}

\item{log.p}{Logical; if TRUE, probabilities are given on the log scale.}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Cumulative Distribution Function of the CTP Distribution
}
\examples{
pctp(0:5, a = 1, b = 0.5, gama = 6)
}
