% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctpfit.R
\name{plot.ctpfit}
\alias{plot.ctpfit}
\title{Plot method for ctpfit objects}
\usage{
\method{plot}{ctpfit}(x, type = c("frequency", "cdf", "qq"), ...)
}
\arguments{
\item{x}{A ctpfit object}

\item{type}{Type of plot: "frequency", "cdf", or "qq"}

\item{...}{Additional graphical parameters}
}
\value{
No return value. Called for its side effect of producing a plot.
}
\description{
Plot method for ctpfit objects
}
