% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.zictpfit}
\alias{print.zictpfit}
\title{Print method for zictpfit objects}
\usage{
\method{print}{zictpfit}(x, ...)
}
\arguments{
\item{x}{A zictpfit object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original fitted model object.
The function is called for its side effect of printing a model summary.
}
\description{
Print method for zictpfit objects
}
