% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzictp.R
\name{pzictp}
\alias{pzictp}
\title{Cumulative Distribution Function of the ZM-CTP Distribution}
\usage{
pzictp(q, a, b, gama, omega, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{a}{Parameter a.}

\item{b}{Parameter b.}

\item{gama}{Parameter gamma.}

\item{omega}{Zero-modification parameter.}

\item{lower.tail}{Logical.}

\item{log.p}{Logical.}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Cumulative Distribution Function of the ZM-CTP Distribution
}
\examples{
pzictp(0:5, a = 1, b = 0.5, gama = 6, omega = 0.3)
}
