% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qctp.R
\name{qctp}
\alias{qctp}
\title{Quantile Function of the CTP Distribution}
\usage{
qctp(p, a, b, gama, lower.tail = TRUE, log.p = FALSE, max_q = 1000)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{a}{Parameter a.}

\item{b}{Parameter b.}

\item{gama}{Parameter gamma.}

\item{lower.tail}{Logical.}

\item{log.p}{Logical.}

\item{max_q}{Maximum x to search.}
}
\value{
Numeric vector of quantiles.
}
\description{
Quantile Function of the CTP Distribution
}
\examples{
qctp(c(0.25, 0.5, 0.75), a = 1, b = 0.5, gama = 6)
}
