% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzictp.R
\name{qzictp}
\alias{qzictp}
\title{Quantile Function of the ZM-CTP Distribution}
\usage{
qzictp(p, a, b, gama, omega, lower.tail = TRUE, log.p = FALSE, max_q = 1000)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{a}{Parameter a.}

\item{b}{Parameter b.}

\item{gama}{Parameter gamma.}

\item{omega}{Zero-modification parameter.}

\item{lower.tail}{Logical.}

\item{log.p}{Logical.}

\item{max_q}{Maximum x to search.}
}
\value{
Numeric vector of quantiles.
}
\description{
Quantile Function of the ZM-CTP Distribution
}
\examples{
qzictp(c(0.25, 0.5, 0.75), a = 1, b = 0.5, gama = 6, omega = 0.3)
}
