% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctp.R
\name{rctp}
\alias{rctp}
\title{Random Generation from the CTP Distribution}
\usage{
rctp(n, a, b, gama)
}
\arguments{
\item{n}{Number of observations.}

\item{a}{Parameter a.}

\item{b}{Parameter b.}

\item{gama}{Parameter gamma.}
}
\value{
Integer vector of random draws.
}
\description{
Random Generation from the CTP Distribution
}
\examples{
set.seed(123)
rctp(10, a = 1, b = 0.5, gama = 6)
}
