% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzictp.R
\name{rzictp}
\alias{rzictp}
\title{Random Generation from the ZM-CTP Distribution}
\usage{
rzictp(n, a, b, gama, omega)
}
\arguments{
\item{n}{Number of observations.}

\item{a}{Parameter a.}

\item{b}{Parameter b.}

\item{gama}{Parameter gamma.}

\item{omega}{Zero-modification parameter.}
}
\value{
Integer vector of random draws.
}
\description{
Random Generation from the ZM-CTP Distribution
}
\examples{
set.seed(123)
rzictp(10, a = 1, b = 0.5, gama = 6, omega = 0.3)
}
