% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmctp-package.R
\docType{package}
\name{zmctp-package}
\alias{zmctp}
\alias{zmctp-package}
\title{zmctp: Zero-Modified Complex Triparametric Pearson Distribution}
\description{
The zmctp package extends the Complex Triparametric Pearson (CTP)
distribution with zero-modified versions for handling overdispersed
count data, particularly when the parameter b approaches zero.
}
\details{
Main functions:
\itemize{
  \item \code{\link{dctp}}, \code{\link{pctp}}, \code{\link{qctp}},
        \code{\link{rctp}} - CTP distribution functions
  \item \code{\link{dzictp}}, \code{\link{pzictp}}, \code{\link{qzictp}},
        \code{\link{rzictp}} - Zero-Modified CTP distribution functions
  \item \code{\link{ctp.fit}} - Fit CTP model
  \item \code{\link{zictp.fit}} - Fit Zero-Modified CTP model
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/roladoja/zmctp}
  \item Report bugs at \url{https://github.com/roladoja/zmctp/issues}
}

}
\author{
\strong{Maintainer}: Rasheedat Oladoja \email{roladoja@ttu.edu} (\href{https://orcid.org/0009-0001-5883-8039}{ORCID})

}
